test fixture code is no longer UI test specific, update the naming
