--- Wikipedia/Code/NetworkFixtureIntercept/UITestHTTPClientProfile.swift
@@ -1,6 +1,0 @@
-/// Shared UI-test network profile values used by the XCTest launch harness and
-/// app-process transport selection.
-enum UITestHTTPClientProfile: String {
-    case e2e = "e2e"
-    case fixtureStrict = "fixture-strict"
-}
