+++ Wikipedia/Code/TestNetworkFixtures/TestHTTPClientProfile.swift
@@ -1,0 +1,6 @@
+/// Shared test network profile values used by the XCTest launch harness and
+/// app-process transport selection.
+enum TestHTTPClientProfile: String {
+    case e2e = "e2e"
+    case fixtureStrict = "fixture-strict"
+}
