+++ Wikipedia/Code/TestNetworkFixtures/TestNetworkFixtureResponse.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 /// Normalized HTTP response returned by the fixture manifest store.
-struct UITestNetworkFixtureResponse {
+struct TestNetworkFixtureResponse {
     let statusCode: Int
     let headers: [String: String]
     let body: Data
