+++ WikipediaUITests/ArticleImageGalleryUITests.swift
@@ -3,7 +3,7 @@ import XCTest
 final class ArticleImageGalleryUITests: XCTestCase {
     func testBohemiaLeadImagePresentsImageGallery() throws {
         try XCTSkipUnless(
-            uiTestConfiguration.httpClientProfile == UITestHTTPClientProfile.e2e.rawValue,
+            uiTestConfiguration.httpClientProfile == TestHTTPClientProfile.e2e.rawValue,
             "Lead-image gallery coverage requires live E2E networking."
         )
