+++ WikipediaUITests/Config/UITestLaunchArgument.swift
@@ -4,7 +4,7 @@ enum UITestLaunchArgument: String {
     case appleLanguages = "-AppleLanguages"
     case didShowOnboarding = "-DidShowOnboarding5.3"
     case hideTipsForTesting = "-WMFHideTipsForTesting"
-    case httpClientProfile = "-WMFUITestHTTPClientProfile"
+    case httpClientProfile = "-WMFTestHTTPClientProfile"
     case resetPreferredLanguages = "-WMFResetPreferredLanguages"
     case suppressReadingChallengeAnnouncement = "-WMFSuppressReadingChallengeAnnouncementForTesting"
     case uiTestLanguageCode = "-WMFUITestLanguageCode"
