+++ WikipediaUITests/README.md
@@ -2,8 +2,8 @@
 This repository uses the Robots pattern for test legibility and organization. Keep this readme aligned with the checked-in workflows, schemes, test plans, and helper APIs.
 ## CI Lanes
 - Pull requests run `.github/workflows/run_ui_tests.yml` against the `WikipediaUITests` scheme and the `English (Light)` configuration from `Test Plans/UITests.xctestplan`.
-- `UITestConfiguration` defaults UI-test launches to fixture mode and forwards `-WMFUITestHTTPClientProfile fixture-strict` to the app.
-- Pull requests also run `.github/workflows/run_e2e_ui_tests.yml` against the same scheme and test plan with the `English (Light, E2E)` configuration, narrowed to the test identifiers listed in `WikipediaUITests/E2ESmokeTests.txt`. That test-plan configuration passes `-WMFUITestHTTPClientProfile e2e` to the UI-test process, so no fixture profile is forwarded to the app and the app uses E2E networking.
+- `UITestConfiguration` defaults UI-test launches to fixture mode and forwards `-WMFTestHTTPClientProfile fixture-strict` to the app.
+- Pull requests also run `.github/workflows/run_e2e_ui_tests.yml` against the same scheme and test plan with the `English (Light, E2E)` configuration, narrowed to the test identifiers listed in `WikipediaUITests/E2ESmokeTests.txt`. That test-plan configuration passes `-WMFTestHTTPClientProfile e2e` to the UI-test process, so no fixture profile is forwarded to the app and the app uses E2E networking.
 - The UI-test workflows publish `.xcresult` bundles as artifacts. Use those bundles for screenshots and failure inspection.
 ## UI Test Robot Pattern
 - Write UI tests as intent-level scripts. Test files should describe the user journey and expected result, not raw selectors, scrolling loops, alert dismissal, or screenshot plumbing.
